+++ src/templates/faktaside/FaktaSideContext.tsx
@@ -1,21 +1,23 @@
 import React, { createContext, ReactNode, useContext } from "react";
-import { FaktasideContext } from "../../../gatsby-utils/createFaktasider";
+import { FaktaSideProps } from "./types";
 
-const inital: FaktasideContext = {
+const inital: FaktaSideProps = {
+  menuData: [],
+  path: "",
   lang: "no",
   id: "N/A",
   slug: "N/A",
   publiseringsTidspunkt: "",
   innhold: [],
   rawData: {},
-  notifikasjoner: [],
+  notifikasjoner: []
 };
 
-const Context = createContext<FaktasideContext>(inital);
+const Context = createContext<FaktaSideProps>(inital);
 
 export const useFaktasideContext = () => useContext(Context);
 
-export const FaktasideProvider = (props: { children: ReactNode } & FaktasideContext) => {
+export const FaktasideProvider = (props: { children: ReactNode } & FaktaSideProps) => {
   const { children, ...rest } = props;
   return <Context.Provider value={rest}>{children}</Context.Provider>;
 };
