+++ src/templates/faktaside/types.ts
@@ -1,0 +1,7 @@
+import { FaktasideContext } from "../../hooks/graphQl/fetchFaktaside";
+import { MenuItem } from "../../hooks/graphQl/menuDataUtils";
+
+export interface FaktaSideProps extends FaktasideContext {
+  path: string;
+  menuData: MenuItem[];
+}
